changed the structure to favor the separation of concerns
